Moved classes and methods related to the push notification to separate class.
